Abstract getting choice field, add notes on DOM
