+++ lib/DOM.md.html
@@ -1,0 +1,63 @@
+# Similarities and differences between basic field DOM structures
+
+_Missing structure of buttons with `a` tags_
+
+(SLIDEBAR) TEXT INPUT
+<label
+        <input
+    <label (error text)
+
+TEXTAREA
+<label
+    <textarea
+===================================
+RADIO/CHECKBOXES/YESNO RADIO
+<label
+    <fieldset
+        <legend
+        <input
+        <label
+            <span
+            <span
+            <span
+            <span
+        <input
+        <label
+            <span
+            <span
+            <span
+            <span
+===================================
+SELECT BOXES
+<label
+        <select
+            <option
+            <option
+
+(COMBOBOX) (choices are same)
+<label
+            <input
+                    <button
+    <select
+        <option
+        <option
+========================
+BUTTON (missing the other button structure (with <a>))
+<fieldset
+    <legend
+        <button
+            <span
+
+
+========================
+NON-INPUTS BEFORE BOTTOM BUTTONS
+<input hidden
+
+NON-INPUTS AT THE VERY BOTTOM (hidden and such)
+<input
+<input
+<input
+<input
+<input
+<input
+<input
